Tool for easing the rpm build process. It's pretty dumb but can analyse a spec file, tar up the source tar and add patch files to the SRPM, build RPMs.
Usage:
  mk_rpms package

There should be a package.spec file available for this to work.
Or a directory called package inside package.spec can be found.

As said: the script is not too smart, the most it does in parsing
variables is to understand "Version: ..." and "Name: ..." lines
in the spec file and correspondingly replace %{version} and %{name}
macros.


git-svn-id: http://svn.oscar.openclustergroup.org/oscar/pkgsrc@4139 18f9d06c-6ae2-0310-8438-938a554feff9