Added support for PackMan smart installer. This is not necessarilly yum specific, so it could be adapted to debian, too. Additional module: PackManSmart.pm.
When installing with the PackManSmart module, the distro detection is
unnecessary. Systeminstaller only needs to be able to use this module (i.e. it
should be able to create a PackMan instance which returns TRUE when the
is_smart() method is invoked. I will probably delete the whole RPMlists and
stagefiles stuff later, when it becomes clear that we have Debian support,
too.


git-svn-id: http://svn.oscar.openclustergroup.org/oscar/pkgsrc@4224 18f9d06c-6ae2-0310-8438-938a554feff9