When updating binary (RPM or SRPM) packages in the OSCAR repository one has two choices: 1: - remove the old package from the repository (svn rm)    - add the new package to the repository      (svn add)    - checkin the new package, commit removal    (svn ci)
2: - move the old package to the new name       (svn mv)
   - copy the new package over the old one      (cp)
   - checkin the new package                    (svn ci)

Method 1 is the obvious one but leads to lost log information. The logs of
past package changes are lost and can only be seen in the log of the
directory.

Method 2 moves the logs with the file, therefore it is much easier to trace
changes. The tools "svn_update" and "svn_update_multiple" are helping you to
use method 2.

Usage examples:

# update yume src.rpm from version 1.2 to 1.4 (the new rpm must exist in the
current directory)
svn_update yume-1.2-1.src.rpm yume-1.4-1.src.rpm

# do the same thing for multiple rpms which have common version strings (for
# example systeminstaller, systemimager, ...
svn_update_multiple 3.5.3-3oscar 3.6.2-1

  ( would updated in packages/sis/distro/common-rpms/ all systemimager
  related rpms).


If you're not convinced yet of the advantage of this method please go to
http://svn.oscar.openclustergroup.org/websvn/listing.php?repname=OSCAR&path=%2Fbranches%2Fbranch-4-2%2Fpackages%2Fsis%2FSRPMS%2F&rev=0&sc=0
and try to find the log history of systeminstaller and of
systemconfigurator. Click on the "View Log" button of the two packages and
you'll see the difference.



git-svn-id: http://svn.oscar.openclustergroup.org/oscar/pkgsrc@4453 18f9d06c-6ae2-0310-8438-938a554feff9